chore: update suitesparse graphblas dependency fix: remove zstd patching

+++ graphblas_sparse_linear_algebra/src/collections/serializer/lz4_serializer.rs
@@ -58,7 +58,7 @@ impl LZ4Serializer {
             || unsafe {
                 GxB_Desc_set(
                     graphblas_descriptor,
-                    GrB_Desc_Field_GxB_COMPRESSION,
+                    GrB_Desc_Field_GxB_COMPRESSION as i32,
                     GxB_COMPRESSION_LZ4,
                 )
             },
