+++ graphblas_sparse_linear_algebra/src/collections/serializer/lz4hc_serializer.rs
@@ -95,7 +95,7 @@ impl LZ4HighCompressionSerializer {
             || unsafe {
                 GxB_Desc_set(
                     graphblas_descriptor,
-                    GrB_Desc_Field_GxB_COMPRESSION,
+                    GrB_Desc_Field_GxB_COMPRESSION as i32,
                     GxB_COMPRESSION_LZ4HC + compression_level.to_graphblas_descriptor_offset(),
                 )
             },
