+++ graphblas_sparse_linear_algebra/src/collections/serializer/serializer_without_compression.rs
@@ -58,7 +58,7 @@ impl SerializerWithoutCompression {
             || unsafe {
                 GxB_Desc_set(
                     graphblas_descriptor,
-                    GrB_Desc_Field_GxB_COMPRESSION,
+                    GrB_Desc_Field_GxB_COMPRESSION as i32,
                     GxB_COMPRESSION_NONE,
                 )
             },
